clean the App.js
