+++ src/App.js
@@ -1,4 +1,3 @@
-import React,{useState ,useEffect }from "react";
 import Register from "./components/users/Registration";
 import ForgotPassword from "./components/users/ForgotPassword";
 import NewPassword from "./components/users/NewPassword";
@@ -12,25 +11,11 @@ import Navbar from "./components/Navbar/Navbar";
 import theme from "./utils/theme";
 import FooterComponent from "./components/FooterComponent/FooterComponent";
 import Workshop from "./components/Workshops/WorkshopBody"
-import axios from 'axios'
 
 
 
 const App = (match) => {
-  const [courses, setCourses] = useState([])
-
-  useEffect(() => {
-    const getCourses = async () => {
-      const response = await axios.get("http://localhost:3001/courses")
-      console.log(response.data)
-      setCourses(response.data)
-    }
-    getCourses()
-  }, [])
-
-
-
-
+  
   return (
     <>
       <Router>
@@ -40,16 +25,20 @@ const App = (match) => {
           <Switch>
             <Route
               exact
+              component={HomeBanner} 
               path="/workshops"
-              render={(props) => 
-              <HomeBanner 
-              courses = {courses}
-              {...props} />}
+           
             />
-            <Route path="/workshops/:workshopId" component={Workshop} 
+
+            <Route 
+            path="/workshops/:workshopId" 
+            component={Workshop} 
             
             />
-            <Route  path="/fees" render={(props) => <Fees {...props} />} />
+            <Route  
+            path="/fees" 
+            render={(props) => <Fees {...props} />} />
+            
             <Route  path="/messages" render={(props) => <Messages {...props} />} />
             <Route path="/login" render={(props) => <Login {...props} />} />
             <Route
