+++ src/components/HomeBanner/HomeBanner.js
@@ -1,8 +1,9 @@
-import React from 'react';
+import React,{useState ,useEffect }from "react";
 import { Container, Hidden, makeStyles, Button } from '@material-ui/core';
 import MainContent from '../MainContent/MainContent';
 import IconsGrid from '../IconsGrid/IconsGrid';
 import background from '../../images/digitalcampus_with_logo.jpg'
+import axios from 'axios'
 
 const useStyles = makeStyles(theme => ({
 
@@ -29,7 +30,20 @@ const useStyles = makeStyles(theme => ({
   },
 }));
 
-const HomeBanner = ({courses,setCourses}) => {
+const HomeBanner = () => {
+
+  const [courses, setCourses] = useState([])
+
+  useEffect(() => {
+    const getCourses = async () => {
+      const response = await axios.get("/courses")
+      console.log(response.data)
+      setCourses(response.data)
+    }
+    getCourses()
+  }, [])
+
+
   const classes = useStyles();
   return (
     <>
